Grammatical error fixed.
In the purpose section, there is a grammatical error. The line is written as, " The goal of this project is to make it much easier to manage jobs on lots of machines and provides high availability." As the 'goal' is the main subject in this sentence so provide must be written in the singular form and not in plural form.
The corrected line could be : "The goal of this project is to make it much easier to manage jobs on lots of machines and provide high availability."

Thank you.